# docs/swagger.yaml
basePath: /api
definitions:
...
  models.Image:
    properties:
      created_at:
        type: string
      id:
        type: integer
      recipe_id:
        type: integer
      updated_at:
        type: string
      url:
        type: string
    type: object
...
  models.Recipe:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/models.Image'
        type: array
      ingredients:
        type: string
      instructions:
        type: string
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.Review:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      recipe_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.ReviewRequest:
    properties:
      content:
        type: string
      recipe_id:
        type: integer
      user_id:
        type: integer
    required:
    - content
    - recipe_id
    - user_id
    type: object
  models.Tag:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      recipes:
        items:
          $ref: '#/definitions/models.Recipe'
        type: array
      updated_at:
        type: string
    type: object
...
host: religious-mitzi-arasycorp-cc5d6628.koyeb.app
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for culinary review API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API Culinary Review
  version: "1.0"
paths:
...
  /api/recipes:
    get:
      description: Retrieves all recipes.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Recipe'
            type: array
      summary: Get all recipes
      tags:
      - recipes
    post:
      consumes:
      - multipart/form-data
      description: Creates a new recipe with the provided details.
      parameters:
      - description: Title of the recipe
        in: formData
        name: title
        required: true
        type: string
      - description: Description of the recipe
        in: formData
        name: description
        required: true
        type: string
      - description: Ingredients of the recipe
        in: formData
        name: ingredients
        required: true
        type: string
      - description: Instructions of the recipe
        in: formData
        name: instructions
        required: true
        type: string
      - description: Images of the recipe
        in: formData
        name: images
        required: true
        type: file
      - description: Tag names in JSON array format
        in: formData
        name: tag_names
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Recipe'
      summary: Create a new recipe
      tags:
      - recipes
  /api/recipes/{id}:
    delete:
      description: Deletes a recipe by ID.
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete a recipe
      tags:
      - recipes
    get:
      description: Retrieves a recipe by its ID.
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipe'
      summary: Get a recipe by ID
      tags:
      - recipes
    put:
      consumes:
      - multipart/form-data
      description: Updates an existing recipe with the provided details.
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: integer
      - description: Title of the recipe
        in: formData
        name: title
        required: true
        type: string
      - description: Description of the recipe
        in: formData
        name: description
        required: true
        type: string
      - description: Ingredients of the recipe
        in: formData
        name: ingredients
        required: true
        type: string
      - description: Instructions of the recipe
        in: formData
        name: instructions
        required: true
        type: string
      - description: Images of the recipe
        in: formData
        name: images
        required: true
        type: file
      - description: Tag names in JSON array format
        in: formData
        name: tag_names
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipe'
      summary: Update an existing recipe
      tags:
      - recipes
...
  /api/reviews:
    get:
      consumes:
      - application/json
      description: Get a list of all reviews
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Review'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all reviews
      tags:
      - reviews
    post:
      consumes:
      - application/json
      description: Create a new review for a recipe
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Review Request
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/models.ReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new review
      tags:
      - reviews
  /api/reviews/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a review by its ID
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete review by ID
      tags:
      - reviews
    get:
      consumes:
      - application/json
      description: Get a review by its ID
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get review by ID
      tags:
      - reviews
    put:
      consumes:
      - application/json
      description: Update an existing review by its ID
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      - description: Review Request
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/models.ReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update review by ID
      tags:
      - reviews
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"